Create User Class (user.py)
User (user.py) is model that defines a user profile. It is the first subclass of the BaseModel superclass. Being a subclass, it inherits all the methods and attributes of BaseModel, so its implementation simply calls the init method of its superclass since User is only required to store four attributes for a user: email, password, first_name and last_name.